# server/firebase/functions/swagger.json
{
    "openapi": "2.0",
    "swagger": "2.0",
    "info": {
        "description": "建立與擷取資料的說明\n目前基於Firebase Cloud Function實作",
        "version": "1.0.0",
        "title": "D_D bookclub API",
        "termsOfService": "",
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        }
    },
    "host": "us-central1-fir-project-85d2e.cloudfunctions.net",
    "basePath": "/app",
    "schemes": "https",
    "paths": {
        "/menber/setup": {
            "post": {
                "tags": ["成員相關"],
                "summary": "新增成員並賦予會員ID",
                "parameters": [{
                        "name": "email",
                        "in": "path",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "name": "password",
                        "in": "path",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "name": "interests",
                        "in": "path",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "name": "neckname",
                        "in": "path",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "name": "occupation",
                        "in": "path",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "name": "introduction",
                        "in": "path",
                        "required": true,
                        "type": "string"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation"
                    },
                    "400": {
                        "description": "Invalid status value"
                    }
                }
            }
        },
        "/menber/quary": {
            "get": {
                "tags": ["成員相關"],
                "summary": "依據會員ID查詢對應資訊",
                "responses": {
                    "200": {
                        "description": "successful operation",
                        "schema": {
                            "$ref": "#/definitions/Menber",
                            "schema": {
                                "$ref": "#/definitions/Menber"
                            }
                        }
                    },
                    "400": {
                        "description": "data{user_id}.{parameter} is not defined"
                    }
                }
            }
        },
        "/interests/add": {
            "post": {
                "tags": ["讀書會類別"],
                "summary": "新增讀書會類別",
                "parameters": [{
                    "name": "name",
                    "description": "欲增加的新類別名稱",
                    "in": "path",
                    "required": true,
                    "type": "string"
                }],
                "responses": {
                    "200": {
                        "description": "successful operation"
                    },
                    "400": {
                        "description": "data{user_id}.{parameter} is not defined"
                    }
                }
            }
        },
        "/interests/delete": {
            "delete": {
                "tags": ["讀書會類別"],
                "summary": "刪除特定的讀書會類別",
                "parameters": [{
                    "name": "name",
                    "description": "欲刪除的新類別名稱",
                    "in": "path",
                    "required": true,
                    "type": "string"
                }],
                "responses": {
                    "200": {
                        "description": "successful operation",
                        "schema": {
                            "$ref": "#/definitions/Category"
                        }
                    },
                    "400": {
                        "description": "data{user_id}.{parameter} is not defined"
                    }
                }
            }
        },
        "/interests/recover": {
            "put": {
                "tags": ["讀書會類別"],
                "summary": "恢復讀書會類別到原先的十種",
                "responses": {
                    "200": {
                        "description": "successful operation",
                        "schema": {
                            "$ref": "#/definitions/Category"
                        }
                    },
                    "400": {
                        "description": "data{user_id}.{parameter} is not defined"
                    }
                }
            }
        },
        "/interests/quary": {
            "get": {
                "tags": ["讀書會類別"],
                "summary": "檢索現有的讀書會類別",
                "responses": {
                    "200": {
                        "description": "successful operation",
                        "schema": {
                            "$ref": "#/definitions/Category"
                        }
                    },
                    "400": {
                        "description": "data{user_id}.{parameter} is not defined"
                    }
                }
            }
        }
    },
    "definitions": {
        "Menber": {
            "type": "object",
            "properties": {
                "email": {
                    "type": "string"
                },
                "interests": {
                    "type": "string"
                },
                "introduction": {
                    "type": "string"
                },
                "neckname": {
                    "type": "string"
                },
                "occupation": {
                    "type": "string"
                },
                "password": {
                    "type": "string"
                }
            }
        },
        "Category": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "integer",
                    "format": "int64"
                },
                "name": {
                    "type": "string"
                }
            },
            "xml": {
                "name": "Category"
            }
        }
    }
}